Bug 705866: Move accelerator files.
Currently, MuPDF puts accelerator files into the directory given
by the environment variables for TEMP, or TMP, or failing that
'/var/tmp' or '/tmp/'.

It has been suggested that this is a security risk because people
can cause access to block by using a fifo.

On windows machines, we try to save in %USERPROFILE%\.config\mupdf.
If USERPROFILE is not defined, we drop back to TEMP and TMP as
before - these are generally user specific on modern Windows boxes.

On other machines, we try $XDG_CACHE_HOME/mupdf, or $HOME/.cache/mupdf.