scripts/wrap/cpp.py: also use pdf_new_foo_*() fns as constructors of fz_* structs.
We previously only considered fz_new_() fns as constructors of fz_* structs,
but the fz_/pdf_ prefix is determined by the args, not the return type, so this
was missing possible constructors.

E.g. this picks up pdf_new_stext_page_from_annot() as a constructor of
fz_stext_page.