Simplify overflow-wrap: break-word handling.
Split the word nodes at each cluster, and flag the fragments as
atomic (can't be broken down further) and overflow-wrap (can be taken
as a wrapping point if in desperate need).

This lets us more easily avoid infinite looping at the beginning
of a line, when even the tiniest split fragments can't fit.

This fixes bug 706481.