html: Don't compensate for box margins twice when sizing images.
Horizontal margins are already taken into account when sizing the
flow box; we can limit the image size to the width of the flox box
directly.

Vertical margins of surrounding blocks also don't matter because they
don't add extra space at the beginning/end of a page break - only the
page margins which are already accounted for in page_h.