Bug 706484: Remove limits on page tree depth.
When we search the page tree for resources, we currently
use a recursive algorithm. To avoid stack overflows we
limit recursive depth to 100.

Use an alternative scheme, suggested by Anatoly Vorobey,
whereby we use a modified version of "Floyd's cycle detection
algorithm".

This does away both with the need for recursion, and for the
need to search an extending list at each step, hence moving
from quadratic time to linear. Very clever!