Bug 706480 followup: Further tweak gridfitting.
The draw device maintains a flag so that it knows when it is
rendering a type3 font. We use this flag to inhibit gridfitting
of images.

If, however, we are using a pattern within a type3 font that
happens to have an image in it, we'd rather NOT inhibit
gridfitting for that.

So, tweak the draw device to clear the type3 flag during
patterns.

This only produces 1 diff in our tests, and that diff is a
cure for exactly the 1 test that I thought was a regression
when the original fix for bug 706480 went in.

tests_private/pdf/PDF_2.0_FTS/fts_23_2311.pdf to pgm @ 200dpi
now has no breakup in the fill of the Type3 d1 triangle glyph.