Fix PDF mark list handling of 0-numbered objects.
In the recent commit for Bug 706506, I incorrectly optimised
pdf_mark_list_push to just exit if called with an object
with num == 0. We didn't check for it, so why store it?

Well, if we don't store it, we have nothing to pop, and hence
we can get out of sync when popping stuff off off the list.

Revert that optimisation here, but keep the actual fix (which
was the memcpy bit).