Attempt to speeedup "copy: chinese-example.pdf" PyMuPDF test.
In order to copy the file, we first open it, then save it.

As part of opening it, we perform tests on each of the
objects in the file. This takes a long time.

This file is strangely constructed in that of the 4 xrefs in
the file, the most basic one is extremely fragmented, making
looking up an object in the xref be almost an O(n) process.

We lookup every object as part of the tests, making this an
O(n^2) process.

Here we move to a process whereby we 'map' the checks across
the objects in the file, moving it back to O(n).