Bug 706703: Avoid stream 'Length' overflow.
When we read 'Length' from the file, we were reading it
as an int. We then 'guessed' the length of the uncompressed
stream by multiplying by 3 - and this overflowed.

Swap to reading it as an int64_t (the best we can do),
converting to a size_t and then working in size_t's
throughout.