Bug 706641: Close diamond and arrow line ending paths.
A regression was introduced in commit
ba307a866adc7d8cd7979b0dafda3b475f4d6e0b when trying to
avoid closing the cloudy border path for e.g. a square
annotation to avoid a straight line from the end of the
last circle to the start of the first circle.

This caused diamond, close arrow and reversed closed arrow
line endings to not be closed. That is fixed in this commit
closing the paths for these three line ending cases.