Bug 705640: Don't always fail early on encrypted EPUB files.
Check individual parts of the file for encryption by parsing out the
encryption.xml file and storing the CipherReference URIs in a list.

For now, we just throw an error whenever we try to open an encrypted
entry in the EPUB. Eventually this can be extended to support decrypting
the data, given access to the appropriate DRM keys.