Treat grey bitmaps correctly for overprint.
When displaying a grey bitmap to CMYK, it should effectively
be treated as a CMYK bitmap with CMY=0 for overprint.

See tests_private/pdf/PDFIA1.7_SUBSET/IA3Z1758.pdf and the
"MD Consult eNews" bitmap.

This appears as black/white when overprint is turned off, but
black/cyan when overprint is turned on.