Bug 706715: Avoid SEGVs in non-color-correct pixmap conversions.
When called with spots, fz_convert_slow_pixmap_samples would assert
and/or SEGV. Here, we update that routine to cope - albeit just
by dropping the spot data.

To do this without a) massively bloating the source, and b) slowing
the existing operation down, we fall back to our standard trick
of templating different optimised implementations.