Journalling: Add pdf_abandon_operation().
Currently, if we get an exception during a pdf_begin_operation()/
pdf_end_operation() phase, we are left with a half-made change in
the journal, requiring gymnastics to remove.

Here, we introduce a cleaner way of handling that, by adding a
pdf_abandon_operation() that does an 'end', but clears any
changes that have been made so far, removing them from the
journal.

So where code might have looked like this before:

--->8----

   pdf_begin_operation();
   fz_try()
      ...
   fz_always()
      pdf_end_operation();
   fz_catch()
      fz_rethrow();
--->8----

it would now look like:

--->8---
   pdf_begin_operation();
   fz_try()
   {
      ...
      pdf_end_operation();
   }
   fz_catch()
   {
      pdf_abandon_operation();
      fz_rethrow();
   }
--->8---

In order for this to work nicely with nested operations, it's
important for us to keep each nestings changes separate from
the previous one until such time as that nesting is closed,
whereupon they can be folded down.

This has meant some internal reworking of journalled changes,
with a new 'pending' list that only gets pulled down onto
the 'current' list when all the nested operations have completed.

The bulk of this commit is moving the code to use begin/end/abandon
rather than just begin/end.