Bug 706769: Fix error given by PDF with huge metatext entries.
Update the display list so that nodes are no longer restricted in
size. If a node is going to be larger than can be signalled in
the node header (9 bits), then send a special value (511) and
follow the node with an (unaligned) size_t with the real size.

This enables us to simplify fz_pack_path too as we no longer
need to pass a 'max' value.

Also, we change FZ_CMD_BEGIN_GROUP to use the colorspace coding
defined within the nodes, rather than using private data space.