Bug 706764: OSS-Fuzz 59600: Fix scaled pixmap weight generation.
Our pixmap scaling relies on us generating tables of 'weights'
for the process to use.

In extreme cases, we can end up so that all the weights we
generate for a pixel are 0. This is such a case.

We used to have a nasty workaround to spot when we were about
to do this, and would fudge one of the weights to be non-zero.
This was being defeated by this particular file.

Here we move to a much nicer, cleaner fudge. If we get to the
end of weight generation for a given destination pixel and we
haven't inserted any weights, insert a single weight for the
central source pixel in the window.