Rework archive internal entrypoints slightly.
Rather than having 'read_entry' and 'open_entry' throw on
failure to open, have them return NULL in such cases.

This allows callers to more easily distinguish "failed due to
the entry not being there" from "failed due to out of memory" or
similar.

This enables us to have new entry points like:

   fz_try_read_archive_entry
   fz_try_parse_xml_archive_entry

and relies on:

   fz_try_open_file
   fz_try_read_file

We could have achieved the same beforehand by calling 'has_entry'
and then 'open_entry', but that isn't atomic (and we might be
operating on directories which could be changing under us).

This enables us to remove some nasty try/catch/discarding in
the "multi" archive handler.