Fix remove-last-char font dropping.
In remove-last-char, we identify the new 'last char on the line'
as being 'prev', and update the line to stop there.

I recently updated this code to drop the font reference too to
avoid leaks. Unfortunately, I was freeing prev->font, when
I should have been freeing prev->next->font - that being the
char that was being dropped!