Continued fixes for joining pdf_journal_entries.
There are 2 places in pdf_end_operation where we join entries.
As such, we need to resolve_undo() both of them, rather than
just one.

Also, in resolve_undo, when we found 2 changes to the same
object, I was keeping the 'before' state from the later change,
which is precisely the wrong thing to do! Fix that.

Also drop the 'before' buffer from later changes to avoid leaks.