Bug 706756: Avoid image filter compression bombs.
The PDF spec allows us to specify a JPEG stream as the source
for a content stream. Effectively this can be a compression
bomb (not least because a JPEG continues to output 'empty'
data to the size that it is specified to be, even if the
underlying data runs out).

So, here, where we open streams for lexing, we refuse to open
streams with compression schemes that can only be images
(JPG, JPX, Fax).