Revert "Bug 706764: OSS-Fuzz 59600: Fix scaled pixmap weight generation."
This reverts commit 8ad9e2393c695e8d8b850a2b735b4a6d02e05a26.

The bugfix introduced a crash in ImageProb2.pdf and ImageProb2y.pdf
in the cluster. Misbehaving on a real file is worse than misbehaving
on a fuzzed file, so for the time being we back this one out until
we have a proper fix that doesn't cause issues for normal files.