Include nested classes in JAR file
In the current configuration, nested classes, e.g. Context$Log.class are not included in the resulting JAR file. This is because files are explicitly included by name, and their name is derived from the existence of .java files, replacing the file ending with .class. However, inner class files like Context$Log.class are generated by the compiler, and don't have corresponding .java files.

Not sure if simply replacing with "com", the root package name, is appropriate ,but it does work for me.