Ensure we can save incrementally when saving to a fz_buffer based output.
Currently the code assumes it is saving to a file (or other stream) that
it can both read, write, truncate and seek.

It needs this to prevent us from clobbering the file we need to copy
data from when appending -- none of which is necessary when writing
to a fz_buffer.

Gate the file content comparison and truncation on having access to
a "verify" stream. If it is absent, only copy the original file
without comparing contents.