gl: Avoid double drop of outline/doc when reloading documents.
If a document was encrypted with the default password, and the user
saved the document with new encryption/password, the saving of the new
file would work fine, but once the file was being reloaded the outline
was dropped an a new password was requested. Once the password was
entered the file was reloaded again, causing the same outline to be
dropped a second time. This commit avoids this by setting the outline
pointer to NULL when the outline is dropped.