Optimize incremental saving by tweaking buffer size.
Increase the size of the buffers used to scan and compare the initial
file contents during an incremental save. This increase in buffer
size speeds up incremental saves up to ten times on large files.

The size was chosen to be 4K empirically on a 64bit Linux running
on an Intel i5 processor.

Any smaller than 4K and the performance is measurably worse,
but increasing it more does not prove beneficial.