Tweak PDF recolor filter.
Filters can be instantiated without giving callbacks for all the
possible rewrite methods.

The processor tracks the ctm throughout processing and passes this
to the image_rewrite method.

We now have an option "repeated_image_rewrite", that, if set true,
will avoid caching images, so every instance of an image will
result in a call to image_rewrite (needed if image_rewrite is
gathering information about the dpis that it is used for).

Also, fix some dictionary leaks.

Also, we were dropping some scn/SCN calls with patterns that use
an array. Unfortunately, in this case we are now setting the
colorspace twice.