End implicit operation also for Popups, Links and Signatures and check boxes.
Previously the implicit operation for these types of annotations were
not ended, which left a stray implicit operation which would never be
ended.

The same issue was present in the code called when toggling checkbox widgets.

Later on if an application such as AppKit, called pdf_undoredo_state(),
it would throw an exception because an operation is ongoing.