Bug 707079: Do not scale color key for indexed images.
The PDF in the bug has an image which has 4 bits per pixel and is
indexed. So the indices are in the range 0..15.

Samples for normal images are scaled to 8 bits per pixel.
Therefore the color key must correspondingly be scaled to the same
8-bit interval so it can be compared to each sample.

The indices for indexed images are not scaled, they are copied
verbatim into each sample byte in the pixmap. Therefore the color
key need not be scaled up/shifted down to the 8-bit interval when
it is compared to each sample in the pixmap.

This commit simply propagates an existing flag indicating whether
the image is indexed and skips scaling of the color key for
indexed images.

But just doing that leaves two regressions in the test cluster.
The images in the testfiles for bug 690786 and 692852 disappear
because they require workarounds similar to those already
implemented in Ghostscript:

 * For 1 bits per component images with indexed colorspaces
   where the minimum boundary of the first masking interval is
   too large, ignore the entire color key masking. See files from
   bugs 690786 and 692852.

 * For 1 bit per component images with indexed colorspaces where
   any masking interval boundary except the minimum boundary
   of the first interval is too large, assume those boundaries
   to be 1.

 * For images with more than 1 bit component where any masking
   interval boundary is too large, mask it with AND down to the
   valid range. See file from bug 707079.

 * For any masking interval boundary that is out of range clamp
   it to the valid range. See files from bugs 689717, 697919 and
   701468.

These workarounds in combination with:

 * Scaling the masking interval boundary up to 8 bits or shifting
   it down to 8 bits. See files from bugs 689717, 697919, 701468
   and 704869.

Fixes the reported issue while keeping all test cases working.