Makefile: add version numbers and installation targets for shared libraries.
* Installation targets install-shared-* build+install C/C++/Python/C# bindings.
* On non-MacOS we append .FZ_VERSION_MINOR.FZ_VERSION_PATCH to shared
  libraries.
* On Linux we create links such as libmupdf.so -> libmupdf.so.23.1 (not
  required on OpenBSD).

For install-shared-* targets we require that USE_SYSTEM_LIBS=yes, otherwise we
fail with a diagnostic.

We install Python mupdf.py and _mupdf.so into location from Python's
sysconfig.get_path('platlib').

In existing calls of ./scripts/mupdfwrap.py:
* Add `--venv` so that we automatically get libclang and swig.
* Use `-d $(OUT)` so we use the right build directory, e.g. if $(build_prefix)
  is set.

$(OUT) is only set correctly (i.e. contains `shared-`) if Make was run with
`shared=yes`. So if $(shared) is not 'yes', rules for shared library targets
that use $(OUT) rerun make with shared=yes.