Fix mutool clean -A and -AA options.
I am removing the guard in pdf_annot_request_resynthesis that forced it
to be a no-op if the document had no changes.

If this breaks anything and the change needs to be reverted, make sure
that pdf_request_(re)synthesis will still work on a freshly opened
document. This is needed for mutool clean's -A and -AA options that
generate missing appearances and force all appearances to be regenerated,
respectively.

For a quick test see if after running
	mutool clean -A annotations_galore_II.pdf
does FreeText annotation 5 have an appearance stream?