Bug 707112: Fix assert in pdfclean to do with objstms.
If we are writing non-incrementally, then we expect only to meet
an 'o' object in our xref if we are saving with objstms. This is
usally ensured by "preloading" the objstms, which has the effect
of renaming 'o' objects to be 'n' objects.

Unfortunately, we were short-circuiting that in the case of
repaired files. Amend the logic a bit here. We don't need to
load the objects in the case of repaired files, but we do need
to rename the type.