Fix another objstm related assert.
Seen when mutool running the following script:

  var doc = Document.openDocument("TestFormExpenses.pdf")
  doc.loadPage(0).getWidgets()[0].setTextValue("Hello")
  doc.save("TestFormExpenses-1.pdf","incremental=yes")

When we promote objects forwards into the increment, they can't be
'o' objects any more.

Also, awkwardly, dowriteobject handles all the objects in range
(whether they need to be written as part of the increment or not),
and so we need to explicitly check whether we should process
the given object.