Rework Freetype locking and memory handling.
Always ensure that we are using the correct context, rather
than the base context.

As this context now always reflects the context that is holding
the lock (or is NULL otherwise), we can use this to spot
fz_drop_font being called recursively when the lock is held.

So, add a new mechanism to the storables whereby we can check
whether a given object can safely be dropped at this point,
and use that to ensure that pdf_drop_font_imp doesn't try to
drop a font while the freetype lock is held by this thread.

Thanks to Sebastian for both finding this problem, and doing
the testing for it!