Always use a specific date when signing on the cluster.
The change in pdf_sign_signature() sets a specific date for use
in the signature appearance.

The change in signer_create_digest() sets a specific date for use
in the signature digest.

This way it is possible to sign and save the same file during two
separate runs in the cluster and get both identical appearance
and file contents.