Another attempt to get adding JBIG2 images working correctly.
If we are including data from a file, then 'embedded' is 0. Just
do what the existing code does.

If we are including data that is embedded within a PDF (or generated
by JBIG2DEC), then embedded = 1.

The old code used to just concatenate the globals stream and the
data stream together, but this causes problems, because the globals
can insist on being kept separately.

For instance, globals as produced by jbig2enc have segments with
segment page association 0. Acrobat checks for these not occurring
in image streams and throws an error otherwise.

The alternative to this would be to rewrite such globals to change
their segment page association.