scripts/wrap/__main__.py: use pipcl.py's support for python build flags.
This fixes finding of python-config on some macos systems.

This commit is a subset of master commit da3c74cd0d3 "scripts/wrap/: fixed
pyodide builds and simplified windows python search.". We haven't cherry-picked
the whole commit because this would require other Pyodide-related changes that
are not suitable for 1.23.x.