LZX decompression: Fix problems seen with wff.chm.
Firstly, fix the bit buffer exhaustion case. This was the
subject of a FIXME from the Rust version.

As part of that we simplify the bit buffer code by removing
the code that attempts to cope with reading more than 16 bits
at a time.

Secondly, we allow for 'empty' huffman trees.