PDF Audit tool.
This commit adds a new tool to mutool, audit.

This scans a PDF file to 'audit' its composition, and produces
a simple html file with various statistics in.

It includes details like how much of the file is used for different
sorts of resources (fonts, forms, bookmarks, images etc), and
on the space usage by different operators within streams.

This is primarily going to be of interest to PDF implementers, or
people looking to optimise the PDF files they create.