Fix libarchive handling after error rework.
A change in the error handling of archives has meant that
slightly grotty handling of libarchive streams has been
exposed.

It seems libarchive can open an archive, only to fail when
we try to read the first header from it. Accordingly, we
have to make fz_is_libarchive_archive try to retrieve the
first header before it's sure it's got a valid libarchive
archive.