Revert 3ca71301c9ded01a5facb06c3edd3b335c01ad86
Do NOT throw an error when stat fails in fz_is_directory.

The accelerator handling, for one, checks for various temp paths
existing using fz_is_directory. With the throw in, this means that
not having a cache directory will fail all opens.

We *could* work around this by changing all the callers to cope
with a more try/catchery, but that's a lot of effort.

Update the declaration so it's clear that we don't throw
if the stat fails.