Repurpose int param in begin_structure device calls.
Previously, the int param given in begin_structure calls was
intended to allow callees to spot the same structure element
being reused - hence it was called 'uid'.

Here we change this slightly. It is now called 'idx' because
it gives the index into the array of sibling structure elements
at that level.

Consider for instance, the following structure tree:

  Book
   |- Document
   |   |- Chapter
   |   |   |- Section
   |   |   `- Section
   |   |- Chapter
   |   |   |- Section
   |   |   `- Section                    <- A
   |   |- Chapter
   |   |   |- Section
   |   |   `- Section
   |   `- Chapter
   |- Document
   |   |- Chapter
   |   |- Chapter
   |   |- Chapter
   |   `- Chapter
   `- Document
       |- Chapter
       |- Chapter
       |- Chapter
       `- Chapter

Point A in this tree can be identified by being nested within
the calls:

 begin_structure(Book, 0)
  begin_structure(Document, 0)
   begin_structure(Chapter, 1)
    begin_structure(Section, 1)

This allows the correct structure to be derived even if the different
elements are (perhaps for graphical composition reasons) sent in
other than purely logical order.

It also allows several pages to be sent and for the document structure
to be correctly carried across between their pages.