If an image is not redacted, do not keep its reference.
The code in pdf_filter_Do_image() checks if the image_filter()
callback returns the same image pointer given as an argument.
If the returned image pointer is not the same care is taken
to always drop the image. If the pointer IS the same then the
image is never dropped. This implies the image filter callbacks
should never keep the image if the same pointer is returned.

For pdf_redact_image_filter_pixels() this means that if redacted
is set then it should return a new image pointer, which it does,
but if redacted is clear, then the incoming image pointer reference
should not be kept.

Without this commit any image that is not redacted will lead to
a memory leak.