Bug 707448: Fix text moving after redaction.
PDF text can (broadly) either be placed using:

 (foo) Tj

or

 [ (foo) 10 (bar) ] TJ

We were adjusting for removed text within a string in the wrong
sense when using the former. This was non-obvious, because the
numbers given in the array in the latter are SUBTRACTED rather
than added to the position, so they are implicitly negated.

Here we recast the code slightly so that the adjustments are made
the same way in either method, and we explicitly negate the
values before writing them to the array.