Reduce document handler open functions to a single one.
This one takes fz_streams for file and accelerator and an
fz_archive for the 'directory context' of a file.

The archive may be NULL, and many document handlers won't
ever use it. Similarly, the streams can be NULL if the
directory itself is supposed to be used to load the
document from (such as an unpacked epub).