Bug 707340: Fix assert in LZXD.
Rather than asserting when we hit a corrupt block (as shown in the
fuzzed examples), update this to be a fz_throw.

Also, move the CHM archive handler definition out of archive.c.
We don't enable it by default, but by having it in archive.c, we
force it to be linked in anyway. We'd rather it dropped out at
link time if it was unused.