Bug 707298: Fix floating point inaccuracy problems in bitmap scaler.
We calculate the left and right edges of the range
"centre +/- window", and pull them "inwards" to the nearest
integers. So right - left can never be larger than 2*window,
right?

Well, sadly, no, cos if centre is sufficiently large, then
the inaccuracies of fp means that left and right might be
further apart that we need.

This will never happen in normal use, and only really happens with
stupidly large scaled images. Add a cheap test to avoid
hitting asserts and overflows.