js: Convert buffer argument to integer, not number.
Previously when one (incorrectly) wrote new Buffer("filename.ext")
murun would try to interpret the string as a number, which ended up
being NaN. This was then cast to a very large integer, which would
trigger ASAN to report an unexpectedly large allocation.

By assuming that the argument is an argument, js_tointeger() will
convert the NaN to 0, thereby no large allocation happens that
bothers ASAN.

Also ensure that the size passed to fz_new_buffer() is always positive.