Makefile: allow control of file modes and venv's.
For `install*` targets, `SO_INSTALL_MODE` will override `install` commands'
default `-m 644` setting for .so files.

For C++, Python and C# builds, `VENV_FLAG` will override the default `--venv`
passed to ./scripts/mupdfwrap.py; so if set to an empty string, the build will
not be done in a venv.

Defaults for `SO_INSTALL_MODE` and `VENV_FLAG` are set using `?=` so they can
be overridden either in the environment or as `make` parameters. For example
`INSTALL_MODE=755 make ...` or `make INSTALL_MODE=755 ...`.