New image redaction option
PDF_REDACT_IMAGE_REMOVE_UNLESS_INVISIBLE is a new option for redaction.
This completely removes images if the image (with clipping applied)
intruddes into the redacted area.

To implement this, the sanitize filter has been updated to keep
track of paths in all circumstances. This, in turn, allows us to
keep a clip rectangle. This is then passed into the image_filter
callback.