Bug 707441: Fix clip handling with alpha.
When we create a clip for a path, we create a mask plane and
draw the path to it.

Then we render the contents of the clip (i.e. the graphical
contents that should be clipped).

Then we pop the clip; we take the contents and simply blend
it through that mask. This phase assumes that any alpha
blending done by the contents of the clip onto the existing
background has already been done.

In order for that alpha blending to have already been done
the destination plane to which the contents were drawn
must have had the contents on.

We were not copying that in all cases - instead, trying to
rely on the presence of an alpha plane to control the blending.
This cannot work without the final clipping operation being
much trickier.

Therefore, move to always copying the contents. This gives various
diffs in the cluster testing, but these appear to be progressions,
including:

 PDF_1.7_FTS/fts_14_1416.pdf (page 1)
 PDF_1.7_FTS/fts_31_3107.pdf (page 2) (significant!)
 PDF_2.0_FTS/fts_31_3128.pdf (page 2) (significant!)