Fix infinite loop seen with PyMuPDF.
The issue is actually a division by zero in fz_dash_lineto.

Consider the following case, drawn in exquisite ascii art:

      +--------+
      |        |           (i.e. b is exactly inline with
      |        |            the bottom of the rectangle,
      +--------+ +b         but off to the right.)
            a+

Following through the code, we take neither of the first
if/else if clauses as a is horizontally on screen.

We take the next 'if', as a is below the screen, and we move
a up to the point at which 'ab' intersects the bottom of the
screen. i.e. a = b. At this point dx = 0.

We then continue through until we reach the bottom ifs, where
dx = 0, against our expectations.

The fix is much simpler than the above explanation. We just
recognise that when dx == 0 and/or dy == 0, we don't need to
move the points.