Fix OSS-Fuzz 66460; assert in fz_dash_lineto().
We perform the operation:

 ax += dx * d/dy     (where dx = bx - ax)

to advance ax towards bx. Unfortunately, due to the vagaries of
limited precision floating point it's possible to have both d and dy
being the same sign, with d < dy, and yet ax might "overrun" bx.

We fix this by using a function to do this 'advance' and checking
for overrun.