Bug 707541: Avoid dereferencing NULL tag name.
The file attached in the bug contains a content.opf that points
to a news.xhtml file that contains no xml-tags, but only the
string "sorry something went wrong". So when xml_to_boxes() calls
fz_xml_root() the resulting root node becomes NULL. This is then
passed to gen2_tag() which mostly could handle a missing tag name,
but the checks for "ol" and "section" were missing a NULL check.