Fix slowdown in stext operation.
SumatraPDF report a problem with stext extraction being much
slower recently.

  https://github.com/sumatrapdfreader/sumatrapdf/issues/4018

Essentially, because of ActualText support, we are being tripped
into decoding images during stext extraction, even when we don't
want to extract images.

The fix suggested by SumatraPDF themselves is incorrect as it
will break the bounds detection for ActualText.

This commit fixes it in a different way. We now modify the bounds
for ActualText without decoding the image, and drop out as soon
as we have done so.