Bug 707455: Correctly account for error indicators.
The attached file has a CSS style that causes a parse error,
specifically the string "IMPORTANT" causes an error here:

margin-left: -1.5em !IMPORTANT;

When throwing an exception about this fz_css_errror() wants to
indicate that the "I" in "IMPORTANT" is where the error occurred,
like so:

margin-left: -1.5em !>I<MPORTANT;

However when populating a local buffer with this information the
string maths forgets to account for the error indcators and the
first character where the error occurred. This means that the
local buffer is overrun and MuPDF built with ASAN complains.

Moreover the first character of the error, "I" in this case,
is repeated after the error indicator due to a missing pointer
post increment:

margin-left: -1.5em !>I<IMPORTANT;

Finally if the string after the error is longer than
PRE_POST_SIZE, fz_css_error() attempts to output a truncated part
of that string, but due to a missing pointer post increment this
is always reduced to "...".

margin-left: -1.5em !>I<...

The fix is to account for the error indicators and the first
character of the error as well as add the two missing increments.