SumatraPDF issue 4018: Introduce pdf_processor 'requirements'.
We add a requirements word to decide whether we need to decode
images for a given PDF processor or not.

Every chained processor needs to set this from it's own
requirements, plus those of it's chained processor.

The run processor can set its requirements from the hints
field of the device.