CFF subsetter: Fix various issues.
The fonts I had tested with originally had not used subr's or
gsubrs, so lots of this code had not been properly exercised.

We were failing to adjust for gsubr and subr numbers being biased.
When we scanned for charstrings, we weren't scanning subrs or
gsubrs either.

We were leaking usage lists.

Our Charset format 2 table code had an off by one error.

When recalculating offsets within the file, we failed to allow
for a subsetted global table.

Local Subr offset is relative to private dict data.

Also, ensure that if we comment out the charstrings subsetting, we
still succeed in writing a valid file. Helpful for debugging.