Bug 707559: When saving old style xref trailer remove XRefStm.
Without this fix the old style xref entries may differ from the ones
in the XRefStm leading to incorrect versions of objects being used.

This is because the when saving a document incrementally the trailer
is copied from the prior version of the document, and only some of the
key-value pairs in the trailer are updated. Notably /XRefStm is never
changed. The old style xref table entries are updated, leading to
differences between the two xref tables.

Since MuPDF never updates /XRefStm contents we can avoid this issue by
simply removing /XRefStm from the trailer when saving documents.