Fix font subsetting.
Font definitions in PDF are heirarchical. Multiple fonts can refer
to the same font file. When we subset the font file, we therefore
need to account for all the things that might refer down to it.

Accordingly, we extend our 'font usage' thing to hold both the
fontfile, and all the top level fonts that point down to it.

We will only perform subsetting if all the top level fonts agree
in terms of it being a cidfont or not.