Fix glitch when printing real numbers.
1) Round properly in pdf_to_int.

The rounding was wrong for negative numbers!
The correct approach to round is floor(x + 0.5).
floor() truncates down, and the float to int implicit
conversion truncates to zero!

This bad rounding changed whole numbers so that -10 became -9!!!

2) Avoid calling pdf_to_int when we already have the number.

The whole reason for checking if the number is an integer is to avoid
doing a floating point conversion, so round tripping through the
pdf_to_int/float machinery again is a waste of CPU cycles.