Bug 707504: Lookup the correct name in the tree node.
Previously we tried to lookup the inherited name value in the
node dictionary, i.e. we were trying to locate the value of the name
/Btn, /Ch, /Sig or /Tx in the node dictionary and this without taking
inherited values into account.

process_sigs() already gets an array of inherited values for /FT in its ft
argument, so the fix is to simply check if this value is /Sig.

Additionally also initialize ft to NULL when pdf_walk_tree() is called.
This ensure that pdf_walk_tree_kid() doesn't memcpy() an uninitialized
value before initializing it when the /FT name is found in a tree node.