Fix redaction problems due to non-nesting of fontname.
In the redaction filter, we keep track of the font descriptor
object in the text state structure. Idiotically, when I added
the font name, I failed to add it to the text state structure
and kept it in the processor state. This meant that it was not
being saved/restored properly over q/Q.

The fix is simply to add it into the structure, with
appropriate refcounting logic.

This means we need a simple 'ref counted string' type, so
add fz_string and the usual keep/droppage to the basic fitz
set of functions.