Fix vanished whitespace in redaction filter.
Previously, whenever we output an operator in the buffer, we'd
have added trailing whitespace. This means that when we come to
add the suffix of the page to mark the redacted areas, we'd know
that we didn't need to insert any whitespace.

Now we are much stingier with the whitespace we output. This means
we will always need to output some whitespace before we add any
extra instructions onto the end of the page.