Bug 707434: Compute offset to bmp pixmap sample as size_t, not int32_t.
When the sizes of bmp images get large enough the computation of the
offset from the beginning of the pixmap samples to the sample that
need to be accessed would overflow, causing the pointer to go beyond
the boundaries of the pixmap samples.

Now we compute the offset as a size_t.