Bug 707560: Tweak text redaction logic.
Previously, when we were asked to see whether a character should
be redacted or not, we found its bbox, intersected that with the
redaction annotation, and took a non-empty result to mean that
there was overlap.

For 'zero size' bboxes, such as diacritics (which sometimes have
an advance of 0), this would mean they could never overlap.

So, we modify the logic here. We still intersect the bbox with the
redaction annotation, but now consider whether the result is a
valid rectangle or not. Invalid rectangles do not overlap. Valid
rectangles (which can include zero area ones) do.