Bug 707545: Don't insert table-cell/table-row boxes in the wrong context.
If a table-cell or table-row element occur outside of a matching
table-row or table context, insert them as plain block elements
instead.

This is not correct according to specifications, which would have us
insert anonymous table and table-row elements to encapsulate these
elements instead.

Add TODO explaining what needs to be done for anonymous table boxes.